Move provider and strategy functions into data.py
